+++ src/index.ts
@@ -2,10 +2,12 @@ import { QlikRepoApi } from "qlik-repo-api";
 import { QlikSaaSApi } from "qlik-saas-api";
 
 import { IRunBookResult, ITaskResult, Runner } from "./RunBook/Runner";
-import { IRunBook } from "./RunBook/RunBook.interfaces";
+import { IRunBook, ITask } from "./RunBook/RunBook.interfaces";
 import { CustomError } from "./util/CustomError";
 import { EventsBus } from "./util/EventBus";
+import { WinOperations } from "./util/WinOperations";
 
+const winOperations = new WinOperations();
 export class Automatiqal {
   private runBook: IRunBook;
   private restInstance: QlikRepoApi.client | QlikSaaSApi.client;
@@ -33,6 +35,11 @@ export class Automatiqal {
       });
     }
 
+    if (runBook.edition == "saas") {
+      console.log("TBA");
+      process.exit(1);
+    }
+
     // initialize the Runner
     this.runner = new Runner(this.runBook, this.restInstance);
   }
@@ -45,7 +52,27 @@ export class Automatiqal {
     if (!this.runBook.tasks || this.runBook.tasks.length == 0)
       throw new CustomError(1000, "RunBook");
 
-    this.checkDuplicateTasks();
+    let errors: string[] = [];
+
+    try {
+      this.checkDuplicateTasks();
+    } catch (e) {
+      errors.push(e.context);
+    }
+
+    try {
+      this.checkWrongOperation();
+    } catch (e) {
+      errors.push(e.context);
+    }
+
+    try {
+      this.checkMissingSource();
+    } catch (e) {
+      errors.push(e.context);
+    }
+
+    if (errors.length > 0) throw new Error(errors.join("\n"));
 
     if (this.runBook.edition != "windows" && this.runBook.edition != "saas")
       throw new CustomError(1001, "RunBook", { arg1: this.runBook.edition });
@@ -53,7 +80,7 @@ export class Automatiqal {
 
   private checkDuplicateTasks(): void {
     const duplicateTasks = this.runBook.tasks
-      .map((t) => t.name)
+      .map((t) => `"${t.name}"`)
       .filter(
         (
           (s) => (v) =>
@@ -64,6 +91,43 @@ export class Automatiqal {
     if (duplicateTasks.length > 0)
       throw new CustomError(1002, "RunBook", {
         arg1: duplicateTasks.join(", "),
+      });
+  }
+
+  private checkMissingSource() {
+    const missingSource = this.runBook.tasks
+      .filter((t) => !t.source && !t.filter)
+      .map((t) => {
+        const i = winOperations.nonSourceOperations.indexOf(t.operation);
+
+        if (i == -1) return t;
+
+        return { name: "EXISTS" } as ITask;
+      })
+      .filter((t) => t.name != "EXISTS")
+      .map((t) => `"${t.name}"`);
+
+    if (missingSource.length > 0)
+      throw new CustomError(1014, "RunBook", {
+        arg1: missingSource.join(", "),
+      });
+  }
+
+  private checkWrongOperation() {
+    const nonExistingOps = this.runBook.tasks
+      .map((t) => {
+        const i = winOperations.names.indexOf(t.operation);
+
+        if (i == -1) return t;
+
+        return { name: "EXISTS" } as ITask;
+      })
+      .filter((t) => t.name != "EXISTS")
+      .map((t) => `"${t.operation}"`);
+
+    if (nonExistingOps.length > 0)
+      throw new CustomError(1013, "RunBook", {
+        arg1: nonExistingOps.join(", "),
       });
   }
 }