+++ src/util/EventBus.ts
@@ -5,15 +5,16 @@ export declare interface EventsBus {
   on(event: "runbook:result", listener: (name: string) => void): this;
   on(event: "runbook:log", listener: (name: string) => void): this;
   on(event: "error", listener: (name: string) => void): this;
+  emit(event: string | symbol, ...args: any[]): boolean;
 }
 
 export class EventsBus extends EventEmitter {
   private static instance: EventsBus;
   constructor() {
+    super();
     if (EventsBus.instance) {
       return EventsBus.instance;
     }
-    super();
     EventsBus.instance = this;
   }
 }