+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.0.5] - 2022-02-03
+
+### Added
+
+- task options are supported (`appendCustomProperties` and `appendTags`)
+- perform checks before run anything:
+  - check for duplicate task names
+  - check if all presented operations actually exists
+  - check if all tasks have filter/source (for operations that require them)
+
+### Fixed
+
+- various small fixes
+
 ## [0.0.4] - 2021-10-XX
 
 ### Fixed